x 1.94 speed up for `ExecutionContext.Run(ec, (o) => { }, null);` where `ec == Default`
x 1.57 speed up for `ExecutionContext.Run(ec, (o) => { }, null);` with `AsyncLocal`s (no notifications)

Results https://github.com/dotnet/coreclr/pull/15629#issuecomment-355708318

Alternative approach to https://github.com/dotnet/coreclr/pull/11100 as suggested by @kouvel https://github.com/dotnet/coreclr/pull/11100#discussion_r113028941

When `Default` context is used in `Run` use it as `null`; which also counts as Default - rather than having two forms of `Default` context internally (Thread starts will null) and saving and restoring between them.

```
                Method |   Current |       New |    Change |
---------------------- |----------:|----------:|----------:|
        Run Default EC | 14.134 ns |  7.303 ns |    x 1.94 |
 Run w/ AsyncLocals EC | 11.986 ns |  7.631 ns |    x 1.57 | 
                 await |  3.152 ns |  3.158 ns |    x 1.00 | 
```

ASP.NET Kestrel Plaintext call counts

**Pre**
![image](https://user-images.githubusercontent.com/1142958/35485603-fbb48a50-0459-11e8-851a-7c1aa3c54564.png)

**Post**
![image](https://user-images.githubusercontent.com/1142958/35485607-09e47914-045a-11e8-950b-9fc04383e2d0.png)


Resolves #11126