@benaadams 2017-12-24
```
Total bytes of diff: -1434 (-0.04% of base)
    diff is an improvement.

Total byte diff includes -184 bytes from reconciling methods
        Base had    8 unique methods,      660 unique bytes
        Diff had    3 unique methods,      476 unique bytes

Top file improvements by size (bytes):
       -1434 : System.Private.CoreLib.dasm (-0.04% of base)

1 total files with size differences (1 improved, 0 regressed), 0 unchanged.

Top method regessions by size (bytes):
         452 : System.Private.CoreLib.dasm - ExecutionContext:Restore(byref,ref) (0/1 methods)
         111 : System.Private.CoreLib.dasm - ExecutionContext:SetLocalValue(ref,ref,bool)
          63 : System.Private.CoreLib.dasm - ExecutionContext:Run(ref,ref,ref)
          12 : System.Private.CoreLib.dasm - Thread:get_ExecutionContext():byref:this (0/1 methods)
          12 : System.Private.CoreLib.dasm - Thread:get_SynchronizationContext():byref:this (0/1 methods)
           2 : System.Private.CoreLib.dasm - SynchronizationContext:SetSynchronizationContext(ref)
           2 : System.Private.CoreLib.dasm - SynchronizationContext:get_Current():ref
           2 : System.Private.CoreLib.dasm - ExecutionContext:Capture():ref

Top method improvements by size (bytes):
       -1438 : System.Private.CoreLib.dasm - AsyncTaskMethodBuilder`1:Start(byref):this (12 methods)
        -384 : System.Private.CoreLib.dasm - ExecutionContext:OnContextChanged(ref,ref) (1/0 methods)
        -123 : System.Private.CoreLib.dasm - ExecutionContext:Restore(ref,ref) (1/0 methods)
         -61 : System.Private.CoreLib.dasm - ExecutionContextSwitcher:Undo(ref):this (1/0 methods)
         -40 : System.Private.CoreLib.dasm - ExecutionContext:EstablishCopyOnWriteScope(ref,byref) (1/0 methods)
         -16 : System.Private.CoreLib.dasm - Thread:set_ExecutionContext(ref):this (1/0 methods)
         -16 : System.Private.CoreLib.dasm - Thread:set_SynchronizationContext(ref):this (1/0 methods)
         -10 : System.Private.CoreLib.dasm - Thread:get_ExecutionContext():ref:this (1/0 methods)

21 total methods with size differences (9 improved, 12 regressed), 16738 unchanged.
```