@benaadams 2017-12-24
x 1.58 speed up for `ExecutionContext.Run(ec, (o) => { }, null);` where `ec == Default`

Pre
```
 Method |      Mean |     Error |    StdDev |
------- |----------:|----------:|----------:|
    Run | 14.690 ns | 0.2907 ns | 0.4261 ns |
```
Post
```
 Method |     Mean |     Error |    StdDev |
------- |---------:|----------:|----------:|
    Run | 9.263 ns | 0.0123 ns | 0.0115 ns |
```