@benaadams 2017-12-25
Corresponding corert change https://github.com/dotnet/coreclr/pull/15629