@benaadams 2017-12-26
Previous call tracing
![image](https://user-images.githubusercontent.com/1142958/34344008-fe137eb0-e9d7-11e7-948e-5f78b3d99875.png)

Post call tracing (`Restore` not invoked)
![image](https://user-images.githubusercontent.com/1142958/34344012-08d400c2-e9d8-11e7-9325-d1ee860ede0e.png)