@stephentoub 2018-01-03
> x 1.58 speed up for ExecutionContext.Run(ec, (o) => { }, null); where ec == Default

Any impact when not using the default?