@benaadams 2018-01-04
> Any impact when not using the default?

Think I have a version that may improve both default (less than current) and non-default (without notifications); setting up some scenarios to test properly.