@benaadams 2018-01-06
Updated PR

Test below + example test in https://github.com/dotnet/coreclr/pull/8216
```csharp
private const int OperationsPerInvoke = 1000;

[Benchmark(OperationsPerInvoke = OperationsPerInvoke)]
public static void Run()
{
    var ec = ExecutionContext.Capture();

    for (var i = 0; i < OperationsPerInvoke; i++)
    {
        ExecutionContext.Run(ec, (o) => { }, null);
    }
}

[Benchmark(OperationsPerInvoke = OperationsPerInvoke)]
public static void RunWithAsyncLocal()
{
    ExecutionContext.Run(ExecutionContext.Capture(), _ =>
    {
        var al = new AsyncLocal<object>();
        al.Value = new object();
        var ec = ExecutionContext.Capture();

        for (var i = 0; i < OperationsPerInvoke; i++)
        {
            ExecutionContext.Run(ec, (o) => { }, null);
        }
    }, null);
}

[Benchmark(OperationsPerInvoke = OperationsPerInvoke)]
public async static Task Await()
{
    for (var i = 0; i < OperationsPerInvoke; i++)
    {
        await Task.CompletedTask;
    }
}
```

Pre
```
            Method |      Mean |     Error |    StdDev |
------------------ |----------:|----------:|----------:|
               Run | 14.134 ns | 0.2580 ns | 0.6230 ns |
 RunWithAsyncLocal | 11.986 ns | 0.2208 ns | 0.2066 ns |
             Await |  3.152 ns | 0.0034 ns | 0.0028 ns |

Reading values
Iters=10000000 Locals= 1 => Time=00:00:00.1094316 GC0=0
Iters=10000000 Locals= 2 => Time=00:00:00.2467233 GC0=0
Iters=10000000 Locals= 4 => Time=00:00:00.5567715 GC0=0
Iters=10000000 Locals= 8 => Time=00:00:01.2746867 GC0=0
Iters=10000000 Locals=16 => Time=00:00:02.9930041 GC0=0
Iters=10000000 Locals=32 => Time=00:00:10.5310053 GC0=0
Writing values
Iters= 1000000 Locals= 1 => Time=00:00:00.0316827 GC0=22
Iters= 1000000 Locals= 2 => Time=00:00:00.0710064 GC0=55
Iters= 1000000 Locals= 4 => Time=00:00:00.2313850 GC0=193
Iters= 1000000 Locals= 8 => Time=00:00:00.5202537 GC0=549
Iters= 1000000 Locals=16 => Time=00:00:01.2778451 GC0=1749
Iters= 1000000 Locals=32 => Time=00:00:33.9625060 GC0=12287
```
Post
```
            Method |     Mean |     Error |    StdDev |
------------------ |---------:|----------:|----------:|
               Run | 8.989 ns | 0.0129 ns | 0.0120 ns |
 RunWithAsyncLocal | 9.639 ns | 0.0120 ns | 0.0112 ns |
             Await | 3.160 ns | 0.0079 ns | 0.0074 ns |

Reading values
Iters=10000000 Locals= 1 => Time=00:00:00.1110434 GC0=0
Iters=10000000 Locals= 2 => Time=00:00:00.2540002 GC0=0
Iters=10000000 Locals= 4 => Time=00:00:00.5741779 GC0=0
Iters=10000000 Locals= 8 => Time=00:00:01.2818136 GC0=0
Iters=10000000 Locals=16 => Time=00:00:03.0224091 GC0=0
Iters=10000000 Locals=32 => Time=00:00:10.4927060 GC0=0
Writing values
Iters= 1000000 Locals= 1 => Time=00:00:00.0314400 GC0=22
Iters= 1000000 Locals= 2 => Time=00:00:00.0732549 GC0=55
Iters= 1000000 Locals= 4 => Time=00:00:00.2325266 GC0=193
Iters= 1000000 Locals= 8 => Time=00:00:00.5284765 GC0=549
Iters= 1000000 Locals=16 => Time=00:00:01.2857983 GC0=1749
Iters= 1000000 Locals=32 => Time=00:00:33.0841431 GC0=12287
```