@benaadams 2018-01-06
Still need to check the effect throwing an exception has on output callstack